proj1/producer/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
kafka-python==2.0.2
requests==2.31.0
urllib3==1.26.6
